+++ PSSDKMan/Commands.ps1
@@ -4,7 +4,7 @@
 	if ($Verbose) { $VerbosePreference = 'Continue' }
 
     if ( !( Test-Path $Global:PSDK_DIR ) ) {
-        Write-Warning "$Global:PSDK_DIR does not exists. Reinitialize posh-gvm"
+        Write-Warning "$Global:PSDK_DIR does not exists. Reinitialize posh-sdk"
         Init-Posh-SDK
     }
 
@@ -179,7 +179,7 @@ function Show-Current-Version($Candidate) {
 function Show-Posh-SDK-Version() {
     $poshGvmVersion = Get-Posh-SDK-Version
     $apiVersion = Get-SDK-API-Version
-    Write-Output "posh-gvm (POwer SHell Groovy enVironment Manager) $poshGvmVersion base on GVM $GVM_BASE_VERSION and GVM API $apiVersion"
+    Write-Output "posh-sdk (POwer SHell Groovy enVironment Manager) $poshGvmVersion base on GVM $GVM_BASE_VERSION and GVM API $apiVersion"
 }
 
 function Show-Broadcast-Message() {